// src/store/modules/user/action-types.ts
/*
 * @version:  ;
 * @description:  ;
 * @autor: Full
 * @date: Do not edit
 */
export enum userActionTypes {
    GET_USREINFO = 'GET_USREINFO',
    GET_USREAMOUNT = 'GET_USREAMOUNT',
    GET_COMBO = 'GET_COMBO',
    GET_PAYLIST = 'GET_PAYLIST',
    GET_PAYAMOUNTLIST = 'GET_PAYAMOUNTLIST',
    GET_RECORDRECHARGE = 'GET_RECORDRECHARGE',
    GET_NEWRECORDRECHARGE = 'GET_NEWRECORDRECHARGE',
    GET_CLEARUSERINFO = 'GET_CLEARUSERINFO',
    GET_WALLETAD = 'GET_WALLETAD',
    GET_SHAREQRDATA = 'GET_SHAREQRDATA',
    GET_SHAREQRDATARECORED = 'GET_SHAREQRDATARECORED',
}